Fixed #19 - Undo/redo properly clear/draw end-of-game status text.
In the Windows GUI (winchen.exe) undoing a move that ended because
of a non-stalemate draw was not clearing the end-of-game text,
and redoing was not detecting end of game.

Both problems were because undo and redo are treated as pseudomoves
by ChessGame::Play(), and that function was incorrectly skipping
the test for a drawn game on receiving pseudomoves.
It was catching stalemates and checkmates (positions with zero legal moves)
but not draws by repetition or 50-move rule.

I'm not sure why I ever thought I should ignore pseudomoves in these
cases but not for other game-ending events.