Stripped out obsolete menu options for piece styles, board size.
There will be a single piece style from now on, so
the menu option to select different ones is obsolete.
Also, for now the board size will be fixed.
If I made the board size variable, it will be done by
dragging the lower right corner to any desired size,
not a menu option. So I removed the menu option.
Also removed the "freshen" menu command; if we ever
need that, it's a bug and needs to be fixed!