Fixed #27 - do not ask for promotion piece when promotion is illegal.
All user interfaces for human players had a flaw:
if promotion was illegal because it would leave the mover
in check, there was still a callback to the UI that asked
what piece to promote to. Now we make sure promotion to
queen is legal before even asking. All 4 promotion pieces
are equally legal; there is no case where one is legal
and the others aren't.