Fixed #28 - display question mark on promotion target square.
When a human player is promoting a pawn, right before displaying
the dialog box to pick the promotion piece, display a pawn
in the promotion square with a little question mark symbol near it.
Before now, we were not even putting the pawn there yet.

This is a much better visual signal to reflect the promotion.