Fixed #35 - fixed crash in tactical benchmark.
The tactical benchmark could crash due to accessing
invalid chess move in ChessBoard::NumberOfRepetitions().
That function tries to repeat making moves on the board,
but the "moves" are not always valid chess moves.

NumberOfRepetitions() already had a safety check to
make sure the board was not edited before doing anything.
The safety check did not work in the case of the benchmark
because the benchmark didn't set an initial FEN
(Forsyth Edwards Notation) string. A non-null FEN string
is required to indicate that a position has been edited.

The fix was to make sure the benchmark initializes
the FEN string for the starting position before
doing each min-max search.