Fixed problem with board editor ending game prematurely.
A previous fix for undo/redo wasn't quite right:

    commit f4e4040d7fcf52a99f5e59c906325313bdd4cb2d
    Author: Don Cross <cosinekitty@gmail.com>
    Date:   Tue Mar 6 13:25:24 2018 -0500

That fix caused the game to end immediately in edit
mode, making it impossible to continue editing!
We still need to prevent undo/redo from getting broken
when the game is over, but every time an edit is made,
we need to suppress detecting the end of the game.